requirements.txt
anyio==3.7.1
appdirs==1.4.4
asgiref==3.5.2
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
boto3==1.28.30
botocore==1.31.30
Brotli==1.1.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==41.0.7
cssselect==1.2.0
dj-database-url==1.2.0
dj-static==0.0.6
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.1
django-storages==1.13.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fake-useragent==1.4.0
google-auth==2.15.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==6.1.1
itsdangerous==2.0.1
Jinja2==3.1.2
jmespath==1.0.1
kaitaistruct==0.10
MarkupSafe==2.1.3
outcome==1.3.0.post0
packaging==23.2
parse==1.20.0
Pillow==9.3.0
pkg_resources==0.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyee==8.2.2
PyJWT==2.6.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
pypng==0.20220715.0
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==1.0.0
pytz==2022.1
requests==2.28.1
requests-html==0.10.0
rsa==4.9
s3transfer==0.6.2
selenium==4.16.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.2
static3==0.7.0
stripe==5.4.0
tqdm==4.66.1
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.7.1
urllib3==1.26.13
w3lib==2.1.2
webdriver-manager==4.0.1
websockets==10.4
Werkzeug==2.3.4
whitenoise==6.5.0
wsproto==1.2.0
zipp==3.15.0
zstandard==0.22.0
